pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.11",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "python-dotenv>=1.0.1",
    "python-socketio[asyncio-client]>=5.12.1",
    "sqlmodel>=0.0.22",
    "uvicorn[standard]>=0.34.0",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "session"
